+++ SandboxiePlus/SandMan/Windows/OptionsWindow.h
@@ -3,6 +3,7 @@
 #include <QtWidgets/QMainWindow>
 #include "ui_OptionsWindow.h"
 #include "SbiePlusAPI.h"
+#include "PendingChanges.h"
 #include "../../MiscHelpers/Common/SettingsWidgets.h"
 
 //////////////////////////////////////////////////////////////////////////
@@ -32,6 +33,7 @@ public:
 		eTemplate,
 		eParent
 	};
+	enum { PendingItemTemplateRole = Qt::UserRole + 103 };
 
 	void LoadCompletionConsent();
 	void SaveCompletionConsent();
@@ -241,7 +243,7 @@ private slots:
 	void OnDelUser();
 	//
 
-	void OnFilterTemplates()		{ ShowTemplates(); }
+	void OnFilterTemplates()		{ FilterTemplates(); }
 	void OnTemplateClicked(QTreeWidgetItem* pItem, int Column);
 	void OnTemplateDoubleClicked(QTreeWidgetItem* pItem, int Column);
 	void OnAddTemplates();
@@ -559,6 +561,7 @@ protected:
 
 	void LoadTemplates();
 	void ShowTemplates();
+	void FilterTemplates();
 	void SaveTemplates();
 	void SetTemplate(const QString& Template, bool bEnabled);
 
@@ -580,6 +583,8 @@ protected:
 	bool m_SkipSaveOnToggle; // Skip saving to config when applying reset settings
 
 	bool m_ConfigDirty;
+	bool m_StartRadioBaselineLoaded;
+	CPendingChanges m_PendingChanges{this, &m_HoldChange, PendingItemTemplateRole, false};
 	QColor m_BorderColor;
 	int m_BorderAlpha;
 	QString m_BoxIcon;